Move Align4 to emath
