+++ crates/egui/src/lib.rs
@@ -458,7 +458,8 @@ pub use epaint::emath;
 pub use ecolor::hex_color;
 pub use ecolor::{Color32, Rgba};
 pub use emath::{
-    lerp, pos2, remap, remap_clamp, vec2, Align, Align2, NumExt, Pos2, Rangef, Rect, Vec2, Vec2b,
+    lerp, pos2, remap, remap_clamp, vec2, Align, Align2, Align4, NumExt, Pos2, Rangef, Rect, Vec2,
+    Vec2b,
 };
 pub use epaint::{
     mutex,
