+++ crates/emath/src/align.rs
@@ -298,3 +298,197 @@ impl std::ops::IndexMut<usize> for Align2 {
 pub fn center_size_in_rect(size: Vec2, frame: Rect) -> Rect {
     Align2::CENTER_CENTER.align_size_within_rect(size, frame)
 }
+
+/// Similar to [`Align2`] but also allows for aligning something outside a rect.
+///
+/// The corner from [`Align4::focus`] on the new rect will be aligned to
+/// the corner from [`Align4::align`] on the original rect.
+///
+/// There are helper constants for the 12 common menu positions:
+/// ```text
+///              ┌───────────┐  ┌────────┐  ┌─────────┐              
+///              │ TOP_START │  │  TOP   │  │ TOP_END │              
+///              └───────────┘  └────────┘  └─────────┘               
+/// ┌──────────┐ ┌────────────────────────────────────┐ ┌───────────┐
+/// │LEFT_START│ │                                    │ │RIGHT_START│
+/// └──────────┘ │                                    │ └───────────┘
+/// ┌──────────┐ │                                    │ ┌───────────┐
+/// │   LEFT   │ │             some_rect              │ │   RIGHT   │
+/// └──────────┘ │                                    │ └───────────┘
+/// ┌──────────┐ │                                    │ ┌───────────┐
+/// │ LEFT_END │ │                                    │ │ RIGHT_END │
+/// └──────────┘ └────────────────────────────────────┘ └───────────┘
+///              ┌────────────┐  ┌──────┐  ┌──────────┐              
+///              │BOTTOM_START│  │BOTTOM│  │BOTTOM_END│              
+///              └────────────┘  └──────┘  └──────────┘              
+/// ```
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct Align4 {
+    pub align: Align2,
+    pub focus: Align2,
+}
+
+impl Default for Align4 {
+    fn default() -> Self {
+        Self::BOTTOM_START
+    }
+}
+
+impl Align4 {
+    pub const TOP_START: Self = Self::new(Align2::LEFT_TOP, Align2::LEFT_BOTTOM);
+    pub const TOP: Self = Self::new(Align2::CENTER_TOP, Align2::CENTER_BOTTOM);
+    pub const TOP_END: Self = Self::new(Align2::RIGHT_TOP, Align2::RIGHT_BOTTOM);
+    pub const RIGHT_START: Self = Self::new(Align2::RIGHT_TOP, Align2::LEFT_TOP);
+    pub const RIGHT: Self = Self::new(Align2::RIGHT_CENTER, Align2::LEFT_CENTER);
+    pub const RIGHT_END: Self = Self::new(Align2::RIGHT_BOTTOM, Align2::LEFT_BOTTOM);
+    pub const BOTTOM_END: Self = Self::new(Align2::RIGHT_BOTTOM, Align2::RIGHT_TOP);
+    pub const BOTTOM: Self = Self::new(Align2::CENTER_BOTTOM, Align2::CENTER_TOP);
+    pub const BOTTOM_START: Self = Self::new(Align2::LEFT_BOTTOM, Align2::LEFT_TOP);
+    pub const LEFT_END: Self = Self::new(Align2::LEFT_BOTTOM, Align2::RIGHT_BOTTOM);
+    pub const LEFT: Self = Self::new(Align2::LEFT_CENTER, Align2::RIGHT_CENTER);
+    pub const LEFT_START: Self = Self::new(Align2::LEFT_TOP, Align2::RIGHT_TOP);
+
+    /// The 12 most common menu positions as an array, for use with [`Align4::find_best_align`].
+    pub const MENU_ALIGNS: [Self; 12] = [
+        Self::BOTTOM_START,
+        Self::BOTTOM_END,
+        Self::TOP_START,
+        Self::TOP_END,
+        Self::RIGHT_END,
+        Self::RIGHT_START,
+        Self::LEFT_END,
+        Self::LEFT_START,
+        // These come last on purpose, we prefer the corner ones
+        Self::TOP,
+        Self::RIGHT,
+        Self::BOTTOM,
+        Self::LEFT,
+    ];
+
+    pub const fn new(align: Align2, focus: Align2) -> Self {
+        Self { align, focus }
+    }
+
+    pub fn align(&self) -> Align2 {
+        self.align
+    }
+
+    pub fn focus(&self) -> Align2 {
+        self.focus
+    }
+
+    /// Convert an [`Align2`] to an [`Align4`], positioning the new rect inside the old one.
+    pub fn from_align2(align: Align2) -> Self {
+        Self {
+            align,
+            focus: align,
+        }
+    }
+
+    /// The center of the new rect will be aligned to a corner of the old rect.
+    pub fn over_corner(align: Align2) -> Self {
+        Self {
+            align,
+            focus: Align2::CENTER_CENTER,
+        }
+    }
+
+    /// Position the new rect outside the old rect.
+    pub fn outside(align: Align2) -> Self {
+        Self {
+            align,
+            focus: align.flip(),
+        }
+    }
+
+    /// Calculate the new rect based on a size and some optional gap.
+    pub fn align_rect(&self, rect: &Rect, size: Vec2, gap: f32) -> Rect {
+        let (pivot, anchor) = self.pivot_pos(rect, gap);
+        pivot.anchor_size(anchor, size)
+    }
+
+    /// Returns a [`Align2`] and a [`Pos2`] that you can e.g. use with [`Area::fixed_pos`]
+    /// and [`Area::pivot`] to align an [`Area`] to some rect.
+    pub fn pivot_pos(&self, rect: &Rect, gap: f32) -> (Align2, Pos2) {
+        (self.focus(), self.anchor(rect, gap))
+    }
+
+    /// Returns a normalized vector that can be used as an offset to create a gap between the rects
+    /// while keeping the edges aligned.
+    pub fn gap_factor(&self) -> Vec2 {
+        let mut gap = -self.focus.to_sign();
+
+        // Align the edges in these cases
+        match *self {
+            Self::TOP_START | Self::TOP_END | Self::BOTTOM_START | Self::BOTTOM_END => {
+                gap.x = 0.0;
+            }
+            Self::LEFT_START | Self::LEFT_END | Self::RIGHT_START | Self::RIGHT_END => {
+                gap.y = 0.0;
+            }
+            _ => {}
+        }
+
+        gap
+    }
+
+    /// Calculator the anchor point for the new rect, based on the old rect and an optional gap.
+    pub fn anchor(&self, rect: &Rect, gap: f32) -> Pos2 {
+        let pos = self.align.pos_in_rect(rect);
+
+        let offset = self.gap_factor() * gap;
+
+        pos + offset
+    }
+
+    /// Returns the 3 alternative [`Align4`]s that are flipped in various ways, for use
+    /// with [`Align4::find_best_align`].
+    pub fn alternatives(self) -> [Self; 3] {
+        let flip_x = Self::new(self.align.flip_x(), self.focus.flip_x());
+        let flip_y = Self::new(self.align.flip_y(), self.focus.flip_y());
+        let flip_xy = Self::new(self.align.flip(), self.focus.flip());
+        [flip_x, flip_y, flip_xy]
+    }
+
+    /// Look for the [`Align4`] that fits best in the available space.
+    ///
+    /// See also:
+    /// - [`Align4::alternatives`] to calculate alternatives
+    /// - [`Align4::MENU_ALIGNS`] for the 12 common menu positions
+    pub fn find_best_align(
+        mut values_to_try: impl Iterator<Item = Self>,
+        available_space: Rect,
+        widget_rect: Rect,
+        gap: f32,
+        size: Vec2,
+    ) -> Self {
+        let area = size.x * size.y;
+
+        let blocked_area = |pos: Self| {
+            let rect = pos.align_rect(&widget_rect, size, gap);
+            area - available_space.intersect(rect).area()
+        };
+
+        let first = values_to_try.next().unwrap_or_default();
+
+        if blocked_area(first) == 0.0 {
+            return first;
+        }
+
+        let mut best_area = blocked_area(first);
+        let mut best = first;
+
+        for align in values_to_try {
+            let blocked = blocked_area(align);
+            if blocked == 0.0 {
+                return align;
+            }
+            if blocked < best_area {
+                best = align;
+                best_area = blocked;
+            }
+        }
+
+        best
+    }
+}
